TaxipassengersBackend/requirements.txt
# Application requirements
Werkzeug==0.16.0
flask==1.0.2
flask-restplus==0.12.1
flask-sqlalchemy==2.3.1
flask-migrate
pytest-flask==0.14.0
freezegun==0.1.11
delorean==1.0.0
faker==1.0.7
pyjwt==1.7.1
cryptography==2.6.1
parse==1.12.0
psycopg2==2.8.2
flake8==3.7.7
requests==2.23.0
celery==4.4.7
flask-cors
redis